Refactor functional tests; fix LogPrinter; RunnerTest is still red
